Changed ALL calls to closedLoopSim.hm (The input file hash map) to eliminate the static reference accross classes.
WARNING!!! This has broken deprecation in the jat/measurements/ directory
- Default constructors now include a HashMap argument which initializes the hash map at object creation rather than initializing statically at instantiation.

Though it makes sense to only have one instance of the input file for the whole simulation, it is difficult to see who has access to the input and how.
If we change the way we handle the input file, each class will see the changes immediately rather than depending on a class which may fall out of use.