I have modified the process models to take a HashMap argument since I noticed a bit late that the initialConditions.txt file was hardcoded there.
There are also a number of other changes in attempts to ascertains some of the differences between JAT and GEONS.

Fixed a bug with UniverseModel.java where the call to set_use_iers(false) was not being properly carried out.  ie other calls reversed it and it was not runtime safe.
It now contains a try-catch which handles the case where the user mistakenly turns on the IERS but does not have the appropriate data file.